Refactor

+++ tests/RootFinder/test_broyden.cc
@@ -31,7 +31,7 @@ TEMPLATE_TEST_CASE("Broyden", "[template]", TEST_VECTOR_FUNCTIONS) {
     SECTION(sol.name()) {
       sol.task(fun.name());
       typename TestType::InputType x_ini, x_out;
-      for (Integer i{0}; i < static_cast<Integer>(fun.guesses().size()); ++i) {
+      for (size_t i{0}; i < fun.guesses().size(); ++i) {
         x_ini = fun.guess(i);
         // Solve without damping
         sol.disable_damped_mode();
